init_sct: add forgotten add_elapsed_time_counter() and init_error_client()
